split out elcid tests, fix the old fixtures, remove unused tests, add in alert investigations add the MicroHaem tag if patients for reasons for investigation is one of the MicoHaems
