+++ elcid/constants.py
@@ -1,0 +1,6 @@
+MICROHAEM_CONSULTATIONS = [
+    "Haematology telephone consult",
+    "Haematology ward round consult",
+]
+
+MICROHAEM_TEAM_NAME = "micro-haem"
