+++ elcid/test/test_models.py
@@ -1,0 +1,214 @@
+"""
+Unittests for the UCLH eLCID OPAL implementation.
+"""
+import datetime
+import ffs
+
+from opal.core import exceptions
+from opal.core.test import OpalTestCase
+from django.test import TestCase
+from opal.models import Patient, Episode, Condition, Synonym
+from elcid.models import Location
+
+HERE = ffs.Path.here()
+TEST_DATA = HERE/'test_data'
+
+
+class AbstractPatientTestCase(TestCase):
+    def setUp(self):
+        super(AbstractPatientTestCase, self).setUp()
+        self.patient = Patient()
+        self.patient.save()
+        self.patient.demographics_set.update(
+            consistency_token="12345678",
+            name="John Smith",
+            hospital_number="AA1111",
+            date_of_birth="1972-06-20",
+        )
+        self.demographics = self.patient.demographics_set.get()
+
+
+class AbstractEpisodeTestCase(AbstractPatientTestCase):
+    def setUp(self):
+        super(AbstractEpisodeTestCase, self).setUp()
+        self.episode = Episode.objects.create(
+            patient=self.patient,
+            consistency_token="12345675"
+        )
+
+
+class DemographicsTest(OpalTestCase, AbstractPatientTestCase):
+
+    def test_to_dict(self):
+        expected_data = {
+            'consistency_token': '12345678',
+            'patient_id': self.patient.id,
+            'id': self.demographics.id,
+            'name': 'John Smith',
+            'created': None,
+            'updated': None,
+            'created_by_id': None,
+            'updated_by_id': None,
+            'date_of_birth': datetime.date(1972, 6, 20),
+            'country_of_birth': '',
+            'country_of_birth_fk_id': None,
+            'country_of_birth_ft': '',
+            'ethnicity': None,
+            'gender': None,
+            'hospital_number': 'AA1111',
+            'nhs_number': None
+            }
+
+        self.assertEqual(expected_data, self.demographics.to_dict(self.user))
+
+    def test_update_from_dict(self):
+        data = {
+            'consistency_token': '12345678',
+            'id': self.demographics.id,
+            'name': 'Johann Schmidt',
+            'date_of_birth': '1972-6-21',
+            'hospital_number': 'AA1112',
+            }
+        self.demographics.update_from_dict(data, self.user)
+        demographics = self.patient.demographics_set.get()
+
+        self.assertEqual('Johann Schmidt', demographics.name)
+        self.assertEqual(datetime.date(1972, 6, 21), demographics.date_of_birth)
+        self.assertEqual('AA1112', demographics.hospital_number)
+
+    def test_update_from_dict_with_missing_consistency_token(self):
+        with self.assertRaises(exceptions.APIError):
+            self.demographics.update_from_dict({}, self.user)
+
+    def test_update_from_dict_with_incorrect_consistency_token(self):
+        with self.assertRaises(exceptions.ConsistencyError):
+            self.demographics.update_from_dict({'consistency_token': '87654321'}, self.user)
+
+
+class LocationTest(OpalTestCase, AbstractEpisodeTestCase):
+
+    def setUp(self):
+        super(LocationTest, self).setUp()
+
+        self.location = Location.objects.create(
+            bed="13",
+            category="Inpatient",
+            consistency_token="12345678",
+            hospital="UCH",
+            ward="T10",
+            episode=self.episode
+        )
+
+    def test_to_dict(self):
+        expected_data = {
+            'consistency_token': '12345678',
+            'episode_id': self.episode.id,
+            'id': self.location.id,
+            'category': 'Inpatient',
+            'hospital': 'UCH',
+            'ward': 'T10',
+            'bed': '13',
+            'created': None,
+            'updated': None,
+            'updated_by_id': None,
+            'created_by_id': None,
+            'opat_discharge': None,
+            'opat_referral': None,
+            'opat_referral_route': None,
+            'opat_referral_team': None,
+            'opat_referral_consultant': None,
+            'opat_referral_team_address': None,
+            }
+        result = {str(k): v for k, v in self.location.to_dict(self.user).iteritems()}
+        self.assertEqual(expected_data, result)
+
+    def test_update_from_dict(self):
+        data = {
+            'consistency_token': '12345678',
+            'id': self.location.id,
+            'category': 'Inpatient',
+            'hospital': 'HH',
+            'ward': 'T10',
+            'bed': '13',
+            }
+        self.location.update_from_dict(data, self.user)
+        self.assertEqual('HH', self.location.hospital)
+
+
+class DiagnosisTest(OpalTestCase, AbstractEpisodeTestCase):
+
+    def setUp(self):
+        super(DiagnosisTest, self).setUp()
+        self.condition_1 = Condition.objects.create(name="Some condition")
+        self.condition_2 = Condition.objects.create(name="Some other condition")
+        Synonym.objects.create(
+            name="Condition synonym",
+            content_object=self.condition_2
+        )
+
+        self.diagnosis = self.episode.diagnosis_set.create(
+            consistency_token="12345678",
+            date_of_diagnosis=datetime.date(2013, 7, 25),
+            details="",
+            provisional=False,
+            condition=self.condition_1.name,
+        )
+
+        self.episode.diagnosis_set.create(
+            condition=self.condition_2.name,
+            date_of_diagnosis=datetime.date(2013, 7, 25),
+            details="",
+            provisional=True,
+        )
+
+        self.diagnosis = self.episode.diagnosis_set.first()
+
+    def test_to_dict(self):
+        expected_data = {
+            'consistency_token': u'12345678',
+            'updated': None,
+            'created': None,
+            'updated_by_id': None,
+            'created_by_id': None,
+            'episode_id': self.episode.id,
+            'id': self.diagnosis.id,
+            'condition': 'Some condition',
+            'condition_fk_id': self.condition_1.id,
+            'condition_ft': u'',
+            'provisional': False,
+            'details': u'',
+            'date_of_diagnosis': datetime.date(2013, 7, 25),
+            }
+
+        result = {str(k): v for k, v in self.diagnosis.to_dict(self.user).iteritems()}
+        self.assertEqual(expected_data, result)
+
+    def test_update_from_dict_with_existing_condition(self):
+        data = {
+            'consistency_token': '12345678',
+            'id': self.diagnosis.id,
+            'condition': 'Some other condition',
+            }
+        self.diagnosis.update_from_dict(data, self.user)
+        diagnosis = self.episode.diagnosis_set.first()
+        self.assertEqual('Some other condition', diagnosis.condition)
+
+    def test_update_from_dict_with_synonym_for_condition(self):
+        data = {
+            'consistency_token': '12345678',
+            'id': self.diagnosis.id,
+            'condition': 'Condition synonym',
+            }
+        self.diagnosis.update_from_dict(data, self.user)
+        diagnosis = self.episode.diagnosis_set.first()
+        self.assertEqual('Some other condition', diagnosis.condition)
+
+    def test_update_from_dict_with_new_condition(self):
+        data = {
+            'consistency_token': '12345678',
+            'id': self.diagnosis.id,
+            'condition': 'New condition',
+            }
+        self.diagnosis.update_from_dict(data, self.user)
+        diagnosis = self.episode.diagnosis_set.first()
+        self.assertEqual('New condition', diagnosis.condition)
